start splitting number tests
